:sparkles: Adapt the ResNet50 model to the task
Create a model inheriting pre-trained weights from ResNet50 trained on
the ImageNet dataset, that is stipped from the last fully-connected
layer replaced with the new linear layer adapted to binary as well as
higher dimensional classification problems.

Model is also storing a hook to gradients from the last convolutional
layer of the original ResNet50 model, that will simplify Grad-CAM
calculations.